refactor(frontend): move native-incompatibility guidance to the Compatibility page
The per-bundle dependency-compare view was the wrong home for the
'this needs a native build' explainer. Move it to the app-level
Compatibility page (/app/:app/compatibility), shown whenever the app has
unresolved incompatibility events. Make the copy generic (no per-bundle
name), keep the Capgo Builder CTA (sell deck) + rollback guidance + the
collapsible 'why native changes need an app-store update' explainer.
Revert the Dependencies page to its original state.
