+++ messages/en.json
@@ -948,13 +948,13 @@
   "compat-verdict-compatible-detail": "This bundle can be delivered over-the-air to devices running {bundle}.",
   "compat-verdict-incompatible": "Not compatible ({count})",
   "compat-verdict-incompatible-detail": "Some packages require an app store update, so this bundle cannot be delivered over-the-air.",
-  "compat-fix-explanation": "Capgo delivers your web and JavaScript code instantly, but it can't change native code. Devices still running the previous native build ({bundle}) may crash or behave incorrectly on this update until your users install a new native build from the app stores.",
+  "compat-fix-explanation": "Capgo delivers your web and JavaScript code instantly, but it can't change native code. When a bundle's native dependencies don't match the native build a device already has, that update may crash or behave incorrectly until your users install a new native build from the app stores.",
   "compat-fix-how-title": "How to fix it",
   "compat-fix-manage-channels": "Manage channels",
   "compat-fix-rebuild-cta": "Rebuild with Capgo Builder",
   "compat-fix-rebuild-detail": "Ship a new native build to the App Store and Play Store so your users' app has the matching native code. Capgo Builder can produce one in minutes — no local Xcode or Android Studio setup needed.",
-  "compat-fix-rollback-detail": "In the meantime, roll your channel back to {bundle} so users stay on the last compatible bundle until the new native build is live.",
-  "compat-fix-title": "This bundle needs a new native build",
+  "compat-fix-rollback-detail": "In the meantime, roll the affected channel back to its last compatible bundle so users stay safe until the new native build is live.",
+  "compat-fix-title": "A new native build is needed",
   "compat-fix-why-detail": "Over-the-air updates can only replace the web layer — JavaScript, HTML and CSS. Native plugins are compiled into the app binary, which only the App Store and Play Store can update. When a bundle's native packages differ from the build installed on a device, its JavaScript can call native APIs that aren't there — which is why it may crash.",
   "compat-fix-why-title": "Why native changes need an app-store update",
   "dependencies": "Dependencies",
